fixed migrations
The old devise migrations dont work anymore with newer versions of devise.

Dropping of video_views and searches does not fail if they do not exist.

+++ db/migrate/20121121134655_drop_video_views_and_searches.rb
@@ -1,6 +1,6 @@
 class DropVideoViewsAndSearches < ActiveRecord::Migration
   def up
-    drop_table :video_views
-    drop_table :searches
+    drop_table :video_views if self.table_exists?(:video_views)
+    drop_table :searches if self.table_exists?(:searches)
   end
 end
